nothing

+++ src/main/java/org/thin/ugc/models/mappers/ContentMapper.java
@@ -5,9 +5,11 @@ import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 import org.thin.ugc.models.Content;
 
+import java.math.BigInteger;
+
 @Mapper
 public interface ContentMapper
 {
     @Select("SELECT * FROM `content` WHERE id = #{id}")
-    Content findByCid(@Param("id") long cid);
+    Content findByCid(@Param("id") BigInteger cid);
 }
